Biddo invamia bid adapters: import common code (#12409)
* Create index.js

* Update biddoBidAdapter.js

* Update invamiaBidAdapter.js
